Add Python 3.14 builds; bump test dependencies to match
This is semi-optimistic concurrency, given 3.14 is still in RC phase. The test
suite ran succcessfully locally (see comment in PR). I synced uv.lock under
Python 3.8 and 3.140rc2, the changes were idempotent.

I went for
- oldest viable looking Maturin (best availability to end-users)
- newest available pytest (benefit from latest fixes in CI jobs)

Maturin 1.8.7 contains a fix for Python 3.14t on Windows.
Maturin 1.9.3 is newest as a I write this.

PyTest 8.4.0 was the first release with official Python 3.14 support.
PyTest 8.4.1 is newest as I write this.

See
- https://www.maturin.rs/changelog.html#187
- https://docs.pytest.org/en/stable/changelog.html#pytest-8-4-1-2025-06-17